par-eval-v3 - extracted parEvalMap* tests to a dedicated Suite
